+++ Editor/SettingsAnalysis/PlayerSettingsAnalyzer.cs
@@ -36,7 +36,7 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
             "Audio: Speaker Mode",
             new[] { Area.BuildSize },
             "<b>UnityEngine.AudioSettings.speakerMode</b> is not set to <b>UnityEngine.AudioSpeakerMode.Mono</b>. The generated build will be larger than necessary.",
-            "To reduce runtime memory consumption of AudioClips change <b>Project Settings ➔ Audio ➔ Default Speaker Mode</b> to <b>Mono</b>. This will reduce half memory usage of stereo AudioClips. It is also recommended considering enabling the <b>Force To Mono</b> AudioClip import setting to reduce import times and build size.")
+            "To reduce runtime memory consumption of AudioClips change <b>Project Settings ➔ Audio ➔ Default Speaker Mode</b> to <b>Mono</b>. This will half memory usage of stereo AudioClips. It is also recommended considering enabling the <b>Force To Mono</b> AudioClip import setting to reduce import times and build size.")
         {
             platforms = new[] { "Android", "iOS"},
             fixer = (issue => {
@@ -96,7 +96,12 @@ namespace Unity.ProjectAuditor.Editor.SettingsAnalysis
 
         internal static void FixSpeakerMode()
         {
-            AudioSettings.speakerMode = AudioSpeakerMode.Mono;
+            AudioConfiguration audioConfiguration = new AudioConfiguration
+            {
+                speakerMode = AudioSpeakerMode.Mono
+            };
+
+            AudioSettings.Reset(audioConfiguration);
         }
     }
 }