+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased]
 * Added _UnityEngine.Object.FindObjectOfType_ usage detection
 * Fixed names of build-generated assets in Build Report 
+* Added Audio Speaker Mode diagnostic fixer
+* Improved Audio Speaker Mode diagnostic recommendations
 * Added AudioSettings speaker mode descriptor and fix to switch it to Mono
 
 ## [0.9.0-preview] - 2022-12-01